Separate Dropdown to subcomponents
