+++ src/components/Dropdown/Dropdown.js
@@ -1,0 +1,30 @@
+// @flow
+import * as React from 'react';
+import { css, cx } from 'emotion';
+import { Button } from '../Button';
+import { IconMore } from '../Icon';
+import { withDropdown } from './withDropdown';
+
+const styles = {
+  wrap: css`
+    position: relative;
+    display: inline-block;
+  `
+}
+
+export const Dropdown = withDropdown(React.forwardRef((
+  {
+    className,
+    children = <Button icon={IconMore} intent='iconic' />,
+    ...props
+  },
+  ref
+) =>
+  <div
+    className={cx(styles.wrap, className)}
+    {...props}
+    ref={ref}
+  >
+    {children}
+  </div>
+));
