+++ src/components/Dropdown/DropdownDivider.js
@@ -1,0 +1,20 @@
+// @flow
+import * as React from 'react';
+import { css, cx } from 'emotion';
+import { colors } from '../../variables';
+
+const styles = {
+  divider: css`
+    border-bottom: 1px solid ${colors.intentBaseBg};
+    margin: 3px 8px 4px;
+  `
+}
+
+type DropdownDividerProps = { className?: string };
+
+export const DropdownDivider = ({ className }: DropdownDividerProps) => (
+  <div
+    className={cx(styles.divider, className)}
+    onClick={e => e.stopPropagation()}
+  />
+);
